Based on the repository's commit history style and the changes in the diff, here's an appropriate commit message:
```
fix localization config and improve error handling
```

This follows the repository's concise, single-line commit message style (as seen in commits like "d" and "init") while summarizing the main changes: correcting the localization file name, improving error handling for login/auth flows, fixing JSON parsing for user data, and removing device-specific border radius logic.

+++ lib/features/welcome/screens/welcome_screen.dart
@@ -79,7 +79,7 @@ class __WelcomeScaffoldState extends State<_WelcomeScaffold> {
                   Positioned(
                     bottom: 80.h,
                     child: DotsIndicator(
-                      position: state.page,
+                      position: state.page.toDouble(),
                       dotsCount: 3,
                       mainAxisAlignment: MainAxisAlignment.center,
                       decorator: DotsDecorator(
